+++ csvkit/utilities/sql2csv.py
@@ -36,10 +36,17 @@ class SQL2CSV(CSVKitUtility):
         )
 
     def main(self):
+        if self.additional_input_expected():
+            self.argparser.error('You must provide an input file or piped data.')
+
         try:
             engine = create_engine(self.args.connection_string)
         except ImportError:
-            raise ImportError('You don\'t appear to have the necessary database backend installed for connection string you\'re trying to use. Available backends include:\n\nPostgresql:\tpip install psycopg2\nMySQL:\t\tpip install MySQL-python\n\nFor details on connection strings and other backends, please see the SQLAlchemy documentation on dialects at: \n\nhttp://www.sqlalchemy.org/docs/dialects/\n\n')
+            raise ImportError("You don't appear to have the necessary database backend installed for connection "
+                              "string you're trying to use. Available backends include:\n\nPostgreSQL:\tpip install "
+                              "psycopg2\nMySQL:\t\tpip install MySQL-python\n\nFor details on connection strings and "
+                              "other backends, please see the SQLAlchemy documentation on dialects at:\n\n"
+                              "http://www.sqlalchemy.org/docs/dialects/\n\n")
 
         connection = engine.connect()