+++ tests/test_utilities/test_csvgrep.py
@@ -42,7 +42,7 @@ class TestCSVGrep(CSVKitTestCase, ColumnsTests, EmptyFileTests, NamesTests):
     def test_match_utf8(self):
         self.assertRows(['-c', '3', '-m', 'ʤ', 'examples/test_utf8.csv'], [
             ['foo', 'bar', 'baz'],
-            ['4' , '5', u'ʤ'],
+            ['4', '5', u'ʤ'],
         ])
 
     def test_no_match(self):
@@ -65,7 +65,7 @@ class TestCSVGrep(CSVKitTestCase, ColumnsTests, EmptyFileTests, NamesTests):
     def test_re_match_utf8(self):
         self.assertRows(['-c', '3', '-r', 'ʤ', 'examples/test_utf8.csv'], [
             ['foo', 'bar', 'baz'],
-            ['4' , '5', u'ʤ'],
+            ['4', '5', u'ʤ'],
         ])
 
     def test_string_match(self):