+++ CHANGELOG.rst
@@ -4,6 +4,7 @@
 Improvements:
 
 * Suppress agate warning about column names not specified when using :code:`--no-header-row`.
+* Prompt the user if additional input is expected (i.e. if no input file or piped data is provided).
 * :doc:`/scripts/csvgrep` adds a :code:`--any-match` (:code:`-a`) flag to select rows where any column matches instead of all columns.
 * :doc:`/scripts/csvjson` no longer emits a property if its value is null.
 * :doc:`/scripts/csvpy` reports an error message if input is provided via STDIN.